requirements.txt
azure-identity>=1.16.0
azure-ai-projects>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
